fix(db): merge migrations into single file
