--- backend/alembic/versions/0c14b62dd6c2_add_module_name_to_transformer.py
@@ -1,30 +1,0 @@
-"""add module_name to transformer
-
-Revision ID: 0c14b62dd6c2
-Revises: 97bbdf50e792
-Create Date: 2025-02-26 12:20:02.104564
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "0c14b62dd6c2"
-down_revision = "97bbdf50e792"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.execute("DELETE FROM transformer")
-    op.add_column("transformer", sa.Column("module_name", sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("transformer", "module_name")
-    # ### end Alembic commands ###
