--- backend/alembic/versions/1172abde1641_add_cascade_delete_to_dag_node.py
@@ -1,84 +1,0 @@
-"""add cascade delete to dag node
-
-Revision ID: 1172abde1641
-Revises: b9bd979b880f
-Create Date: 2025-02-24 13:04:59.869798
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "1172abde1641"
-down_revision = "b9bd979b880f"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Add cascade delete to dag_node.dag_id
-    op.drop_constraint("dag_node_dag_id_fkey", "dag_node", type_="foreignkey")
-    op.create_foreign_key(
-        "fk_dag_node_dag_id",
-        "dag_node",
-        "sync_dag_definition",
-        ["dag_id"],
-        ["id"],
-        ondelete="CASCADE",
-    )
-
-    # Add cascade delete to dag_edge.dag_id
-    op.drop_constraint("dag_edge_dag_id_fkey", "dag_edge", type_="foreignkey")
-    op.create_foreign_key(
-        "fk_dag_edge_dag_id",
-        "dag_edge",
-        "sync_dag_definition",
-        ["dag_id"],
-        ["id"],
-        ondelete="CASCADE",
-    )
-
-    # Add cascade delete to dag_edge node references
-    op.drop_constraint("dag_edge_from_node_id_fkey", "dag_edge", type_="foreignkey")
-    op.drop_constraint("dag_edge_to_node_id_fkey", "dag_edge", type_="foreignkey")
-    op.create_foreign_key(
-        "fk_dag_edge_from_node_id",
-        "dag_edge",
-        "dag_node",
-        ["from_node_id"],
-        ["id"],
-        ondelete="CASCADE",
-    )
-    op.create_foreign_key(
-        "fk_dag_edge_to_node_id", "dag_edge", "dag_node", ["to_node_id"], ["id"], ondelete="CASCADE"
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    # Remove cascade delete from dag_edge node references
-    op.drop_constraint("fk_dag_edge_to_node_id", "dag_edge", type_="foreignkey")
-    op.drop_constraint("fk_dag_edge_from_node_id", "dag_edge", type_="foreignkey")
-    op.create_foreign_key(
-        "dag_edge_to_node_id_fkey", "dag_edge", "dag_node", ["to_node_id"], ["id"]
-    )
-    op.create_foreign_key(
-        "dag_edge_from_node_id_fkey", "dag_edge", "dag_node", ["from_node_id"], ["id"]
-    )
-
-    # Remove cascade delete from dag_edge.dag_id
-    op.drop_constraint("fk_dag_edge_dag_id", "dag_edge", type_="foreignkey")
-    op.create_foreign_key(
-        "dag_edge_dag_id_fkey", "dag_edge", "sync_dag_definition", ["dag_id"], ["id"]
-    )
-
-    # Remove cascade delete from dag_node.dag_id
-    op.drop_constraint("fk_dag_node_dag_id", "dag_node", type_="foreignkey")
-    op.create_foreign_key(
-        "dag_node_dag_id_fkey", "dag_node", "sync_dag_definition", ["dag_id"], ["id"]
-    )
-    # ### end Alembic commands ###
