--- backend/alembic/versions/1ebb07e18a07_make_sync_point_at_connection_ids.py
@@ -1,50 +1,0 @@
-"""Make sync point at connection ids
-
-Revision ID: 1ebb07e18a07
-Revises: 0ebb599fc7fd
-Create Date: 2025-01-02 23:19:10.936354
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '1ebb07e18a07'
-down_revision = '0ebb599fc7fd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('sync', sa.Column('source_connection_id', sa.UUID(), nullable=False))
-    op.add_column('sync', sa.Column('destination_connection_id', sa.UUID(), nullable=True))
-    op.add_column('sync', sa.Column('embedding_model_connection_id', sa.UUID(), nullable=True))
-    op.drop_constraint('sync_destination_integration_credential_id_fkey', 'sync', type_='foreignkey')
-    op.drop_constraint('sync_source_integration_credential_id_fkey', 'sync', type_='foreignkey')
-    op.drop_constraint('sync_embedding_model_integration_credential_id_fkey', 'sync', type_='foreignkey')
-    op.create_foreign_key(None, 'sync', 'connection', ['source_connection_id'], ['id'])
-    op.create_foreign_key(None, 'sync', 'connection', ['embedding_model_connection_id'], ['id'])
-    op.create_foreign_key(None, 'sync', 'connection', ['destination_connection_id'], ['id'])
-    op.drop_column('sync', 'destination_integration_credential_id')
-    op.drop_column('sync', 'embedding_model_integration_credential_id')
-    op.drop_column('sync', 'source_integration_credential_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('sync', sa.Column('source_integration_credential_id', sa.UUID(), autoincrement=False, nullable=False))
-    op.add_column('sync', sa.Column('embedding_model_integration_credential_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.add_column('sync', sa.Column('destination_integration_credential_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.drop_constraint(None, 'sync', type_='foreignkey')
-    op.drop_constraint(None, 'sync', type_='foreignkey')
-    op.drop_constraint(None, 'sync', type_='foreignkey')
-    op.create_foreign_key('sync_embedding_model_integration_credential_id_fkey', 'sync', 'integration_credential', ['embedding_model_integration_credential_id'], ['id'])
-    op.create_foreign_key('sync_source_integration_credential_id_fkey', 'sync', 'integration_credential', ['source_integration_credential_id'], ['id'])
-    op.create_foreign_key('sync_destination_integration_credential_id_fkey', 'sync', 'integration_credential', ['destination_integration_credential_id'], ['id'])
-    op.drop_column('sync', 'embedding_model_connection_id')
-    op.drop_column('sync', 'destination_connection_id')
-    op.drop_column('sync', 'source_connection_id')
-    # ### end Alembic commands ###
