--- backend/alembic/versions/1fb5731c172e_make_white_label_point_at_source_.py
@@ -1,32 +1,0 @@
-"""Make white label point at source through short name
-
-Revision ID: 1fb5731c172e
-Revises: 9db6375e684f
-Create Date: 2025-02-05 10:25:55.936240
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '1fb5731c172e'
-down_revision = '9db6375e684f'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('white_label', sa.Column('source_short_name', sa.String(), nullable=False))
-    op.create_foreign_key(None, 'white_label', 'source', ['source_short_name'], ['short_name'])
-    op.drop_column('white_label', 'source_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('white_label', sa.Column('source_id', sa.UUID(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'white_label', type_='foreignkey')
-    op.drop_column('white_label', 'source_short_name')
-    # ### end Alembic commands ###
