--- backend/alembic/versions/52e0072a0691_change_dag_node_fields.py
@@ -1,38 +1,0 @@
-"""change dag node fields
-
-Revision ID: 52e0072a0691
-Revises: 622b6c7a7dfd
-Create Date: 2025-02-17 16:30:39.571645
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '52e0072a0691'
-down_revision = '622b6c7a7dfd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('dag_node_destination_id_fkey', 'dag_node', type_='foreignkey')
-    op.drop_constraint('dag_node_transformer_id_fkey', 'dag_node', type_='foreignkey')
-    op.drop_constraint('dag_node_source_id_fkey', 'dag_node', type_='foreignkey')
-    op.drop_column('dag_node', 'source_id')
-    op.drop_column('dag_node', 'destination_id')
-    op.drop_column('dag_node', 'transformer_id')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('dag_node', sa.Column('transformer_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.add_column('dag_node', sa.Column('destination_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.add_column('dag_node', sa.Column('source_id', sa.UUID(), autoincrement=False, nullable=True))
-    op.create_foreign_key('dag_node_source_id_fkey', 'dag_node', 'source', ['source_id'], ['id'])
-    op.create_foreign_key('dag_node_transformer_id_fkey', 'dag_node', 'transformer', ['transformer_id'], ['id'])
-    op.create_foreign_key('dag_node_destination_id_fkey', 'dag_node', 'destination', ['destination_id'], ['id'])
-    # ### end Alembic commands ###
