--- backend/alembic/versions/6836af24ebf3_add_embedding_models_sources_.py
@@ -1,119 +1,0 @@
-"""Add embedding models sources destinations tables
-
-Revision ID: 6836af24ebf3
-Revises: c7ba645ff3e0
-Create Date: 2024-12-27 17:11:24.776647
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = '6836af24ebf3'
-down_revision = 'c7ba645ff3e0'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('destination',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('short_name', sa.String(), nullable=False),
-    sa.Column('class_name', sa.String(), nullable=False),
-    sa.Column('auth_config_class', sa.String(), nullable=True),
-    sa.Column('auth_type', sa.Enum('oauth2', 'oauth2_with_refresh', 'oauth2_with_refresh_rotating', 'trello_auth', 'api_key', 'native_functionality', 'config_class', name='authtype'), nullable=True),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('short_name')
-    )
-    op.create_table('embedding_model',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('short_name', sa.String(), nullable=False),
-    sa.Column('provider', sa.String(), nullable=False),
-    sa.Column('auth_config_class', sa.String(), nullable=True),
-    sa.Column('auth_type', sa.Enum('oauth2', 'oauth2_with_refresh', 'oauth2_with_refresh_rotating', 'trello_auth', 'api_key', 'native_functionality', 'config_class', name='authtype'), nullable=True),
-    sa.Column('model_name', sa.String(), nullable=True),
-    sa.Column('model_version', sa.String(), nullable=True),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('short_name')
-    )
-    op.create_table('source',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('short_name', sa.String(), nullable=False),
-    sa.Column('class_name', sa.String(), nullable=False),
-    sa.Column('auth_config_class', sa.String(), nullable=True),
-    sa.Column('auth_type', sa.Enum('oauth2', 'oauth2_with_refresh', 'oauth2_with_refresh_rotating', 'trello_auth', 'api_key', 'native_functionality', 'config_class', name='authtype'), nullable=True),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name'),
-    sa.UniqueConstraint('short_name')
-    )
-    op.create_table('integration_credential',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('integration_short_name', sa.String(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('integration_type', sa.String(), nullable=False),
-    sa.Column('auth_credential_type', sa.String(), nullable=False),
-    sa.Column('encrypted_credentials', sa.Text(), nullable=False),
-    sa.Column('auth_config_class', sa.String(), nullable=False),
-    sa.Column('source_id', sa.UUID(), nullable=True),
-    sa.Column('destination_id', sa.UUID(), nullable=True),
-    sa.Column('embedding_model_id', sa.UUID(), nullable=True),
-    sa.Column('organization_id', sa.UUID(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.Column('created_by_email', sa.String(), nullable=False),
-    sa.Column('modified_by_email', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['created_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['destination_id'], ['destination.id'], ),
-    sa.ForeignKeyConstraint(['embedding_model_id'], ['embedding_model.id'], ),
-    sa.ForeignKeyConstraint(['modified_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['organization_id'], ['organization.id'], ),
-    sa.ForeignKeyConstraint(['source_id'], ['source.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('sync',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('schedule', sa.String(length=100), nullable=False),
-    sa.Column('source_integration_credential_id', sa.UUID(), nullable=False),
-    sa.Column('destination_integration_credential_id', sa.UUID(), nullable=True),
-    sa.Column('embedding_model_integration_credential_id', sa.UUID(), nullable=True),
-    sa.Column('organization_id', sa.UUID(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.Column('created_by_email', sa.String(), nullable=False),
-    sa.Column('modified_by_email', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['created_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['destination_integration_credential_id'], ['integration_credential.id'], ),
-    sa.ForeignKeyConstraint(['embedding_model_integration_credential_id'], ['integration_credential.id'], ),
-    sa.ForeignKeyConstraint(['modified_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['organization_id'], ['organization.id'], ),
-    sa.ForeignKeyConstraint(['source_integration_credential_id'], ['integration_credential.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('sync')
-    op.drop_table('integration_credential')
-    op.drop_table('source')
-    op.drop_table('embedding_model')
-    op.drop_table('destination')
-    op.execute("DROP TYPE IF EXISTS authtype")
-    # ### end Alembic commands ###
