--- backend/alembic/versions/95e59aff8b7a_add_module_name_and_class_name_to_.py
@@ -1,31 +1,0 @@
-"""add module name and class name to entity definition
-
-Revision ID: 95e59aff8b7a
-Revises: 0c14b62dd6c2
-Create Date: 2025-02-26 16:17:23.184247
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "95e59aff8b7a"
-down_revision = "0c14b62dd6c2"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("entity_definition", sa.Column("module_name", sa.String(), nullable=False))
-    op.add_column("entity_definition", sa.Column("class_name", sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("entity_definition", "class_name")
-    op.drop_column("entity_definition", "module_name")
-    # ### end Alembic commands ###
