--- backend/alembic/versions/97bbdf50e792_add_method_name_to_transformer.py
@@ -1,28 +1,0 @@
-"""add method_name to transformer
-
-Revision ID: 97bbdf50e792
-Revises: 9832774ae89f
-Create Date: 2025-02-24 17:01:28.529349
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '97bbdf50e792'
-down_revision = '9832774ae89f'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('transformer', sa.Column('method_name', sa.String(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('transformer', 'method_name')
-    # ### end Alembic commands ###
