--- backend/alembic/versions/9832774ae89f_add_transformer_id_to_dag_node.py
@@ -1,30 +1,0 @@
-"""add transformer id to dag node
-
-Revision ID: 9832774ae89f
-Revises: 3a3c65856c56
-Create Date: 2025-02-24 16:19:57.794596
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '9832774ae89f'
-down_revision = '3a3c65856c56'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('dag_node', sa.Column('transformer_id', sa.UUID(), nullable=True))
-    op.create_foreign_key(None, 'dag_node', 'transformer', ['transformer_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'dag_node', type_='foreignkey')
-    op.drop_column('dag_node', 'transformer_id')
-    # ### end Alembic commands ###
