--- backend/alembic/versions/a734ec1886d4_rework_relations_between_connection_and_.py
@@ -1,75 +1,0 @@
-"""Rework relations between connection and source, destination and embeddingmodel
-
-Revision ID: a734ec1886d4
-Revises: a79d4ff146b2
-Create Date: 2025-01-22 13:00:52.299529
-
-"""
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "a734ec1886d4"
-down_revision = "a79d4ff146b2"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("entity_sync_job_id_fkey", "entity", type_="foreignkey")
-    op.drop_constraint("entity_sync_id_fkey", "entity", type_="foreignkey")
-    op.create_foreign_key(
-        "fk_entity_sync_id", "entity", "sync", ["sync_id"], ["id"], ondelete="CASCADE"
-    )
-    op.create_foreign_key(
-        "fk_entity_sync_job_id", "entity", "sync_job", ["sync_job_id"], ["id"], ondelete="CASCADE"
-    )
-    op.add_column("connection", sa.Column("short_name", sa.String(), nullable=False))
-    op.drop_constraint("connection_source_id_fkey", "connection", type_="foreignkey")
-    op.drop_constraint("connection_embedding_model_id_fkey", "connection", type_="foreignkey")
-    op.drop_constraint("connection_destination_id_fkey", "connection", type_="foreignkey")
-    op.drop_column("connection", "embedding_model_id")
-    op.drop_column("connection", "source_id")
-    op.drop_column("connection", "destination_id")
-    op.drop_constraint("sync_job_sync_id_fkey", "sync_job", type_="foreignkey")
-    op.create_foreign_key(
-        "fk_sync_job_sync_id", "sync_job", "sync", ["sync_id"], ["id"], ondelete="CASCADE"
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("fk_sync_job_sync_id", "sync_job", type_="foreignkey")
-    op.create_foreign_key("sync_job_sync_id_fkey", "sync_job", "sync", ["sync_id"], ["id"])
-    op.add_column(
-        "connection", sa.Column("destination_id", sa.UUID(), autoincrement=False, nullable=True)
-    )
-    op.add_column(
-        "connection", sa.Column("source_id", sa.UUID(), autoincrement=False, nullable=True)
-    )
-    op.add_column(
-        "connection", sa.Column("embedding_model_id", sa.UUID(), autoincrement=False, nullable=True)
-    )
-    op.create_foreign_key(
-        "connection_destination_id_fkey", "connection", "destination", ["destination_id"], ["id"]
-    )
-    op.create_foreign_key(
-        "connection_embedding_model_id_fkey",
-        "connection",
-        "embedding_model",
-        ["embedding_model_id"],
-        ["id"],
-    )
-    op.create_foreign_key(
-        "connection_source_id_fkey", "connection", "source", ["source_id"], ["id"]
-    )
-    op.drop_column("connection", "short_name")
-    op.drop_constraint("fk_entity_sync_job_id", "entity", type_="foreignkey")
-    op.drop_constraint("fk_entity_sync_id", "entity", type_="foreignkey")
-    op.create_foreign_key("entity_sync_id_fkey", "entity", "sync", ["sync_id"], ["id"])
-    op.create_foreign_key("entity_sync_job_id_fkey", "entity", "sync_job", ["sync_job_id"], ["id"])
-    # ### end Alembic commands ###
