--- backend/alembic/versions/a79d4ff146b2_change_white_label_source_id_type.py
@@ -1,43 +1,0 @@
-"""Change white label source id type
-
-Revision ID: a79d4ff146b2
-Revises: fa3b5e05152a
-Create Date: 2025-01-20 14:09:46.194331
-
-"""
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "a79d4ff146b2"
-down_revision = "fa3b5e05152a"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "white_label",
-        "source_id",
-        existing_type=sa.VARCHAR(),
-        type_=sa.UUID(),
-        postgresql_using="source_id::uuid",
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "white_label",
-        "source_id",
-        existing_type=sa.UUID(),
-        type_=sa.VARCHAR(),
-        postgresql_using="source_id::varchar",
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
