--- backend/alembic/versions/b9bd979b880f_add_cascade_delete_to_sync_dag_.py
@@ -1,42 +1,0 @@
-"""add cascade delete to sync dag definition
-
-Revision ID: b9bd979b880f
-Revises: 3a3c65856c56
-Create Date: 2025-02-24 12:56:21.944929
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "b9bd979b880f"
-down_revision = "3a3c65856c56"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(
-        "sync_dag_definition_sync_id_fkey", "sync_dag_definition", type_="foreignkey"
-    )
-    op.create_foreign_key(
-        "fk_sync_dag_definition_sync_id",
-        "sync_dag_definition",
-        "sync",
-        ["sync_id"],
-        ["id"],
-        ondelete="CASCADE",
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("fk_sync_dag_definition_sync_id", "sync_dag_definition", type_="foreignkey")
-    op.create_foreign_key(
-        "sync_dag_definition_sync_id_fkey", "sync_dag_definition", "sync", ["sync_id"], ["id"]
-    )
-    # ### end Alembic commands ###
