--- backend/alembic/versions/c7ba645ff3e0_initial_migration.py
@@ -1,68 +1,0 @@
-"""Initial migration.
-
-Revision ID: c7ba645ff3e0
-Revises:
-Create Date: 2024-12-24 14:36:21.870910
-
-"""
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = 'c7ba645ff3e0'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('organization',
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name')
-    )
-    op.create_table('user',
-    sa.Column('full_name', sa.String(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('password', sa.String(), nullable=False),
-    sa.Column('is_active', sa.Boolean(), nullable=False),
-    sa.Column('is_superuser', sa.Boolean(), nullable=False),
-    sa.Column('organization_id', sa.UUID(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['organization_id'], ['organization.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('email')
-    )
-    op.create_table('api_key',
-    sa.Column('key', sa.String(), nullable=False),
-    sa.Column('key_prefix', sa.String(length=8), nullable=False),
-    sa.Column('expiration_date', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('organization_id', sa.UUID(), nullable=False),
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.Column('modified_at', sa.DateTime(), nullable=False),
-    sa.Column('created_by_email', sa.String(), nullable=False),
-    sa.Column('modified_by_email', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['created_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['modified_by_email'], ['user.email'], ),
-    sa.ForeignKeyConstraint(['organization_id'], ['organization.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('key')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('api_key')
-    op.drop_table('user')
-    op.drop_table('organization')
-    # ### end Alembic commands ###
