--- backend/alembic/versions/d3969b3eb198_add_chat_models.py
@@ -1,91 +1,0 @@
-"""add_chat_models
-
-Revision ID: d3969b3eb198
-Revises: e0af9e9a55fe
-Create Date: 2025-01-30 13:52:51.729766
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "d3969b3eb198"
-down_revision = "e0af9e9a55fe"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "chat",
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column("sync_id", sa.UUID(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column("model_name", sa.String(length=50), nullable=False),
-        sa.Column("model_settings", sa.JSON(), nullable=False),
-        sa.Column("search_settings", sa.JSON(), nullable=False),
-        sa.Column("organization_id", sa.UUID(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("modified_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by_email", sa.String(), nullable=False),
-        sa.Column("modified_by_email", sa.String(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["sync_id"],
-            ["sync.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["created_by_email"],
-            ["user.email"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["modified_by_email"],
-            ["user.email"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["organization_id"],
-            ["organization.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "chat_message",
-        sa.Column("chat_id", sa.UUID(), nullable=False),
-        sa.Column("content", sa.Text(), nullable=False),
-        sa.Column("role", sa.Enum("USER", "ASSISTANT", "SYSTEM", name="chatrole"), nullable=False),
-        sa.Column("tokens_used", sa.Integer(), nullable=True),
-        sa.Column("organization_id", sa.UUID(), nullable=False),
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("modified_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by_email", sa.String(), nullable=False),
-        sa.Column("modified_by_email", sa.String(), nullable=False),
-        sa.ForeignKeyConstraint(["chat_id"], ["chat.id"], ondelete="CASCADE"),
-        sa.ForeignKeyConstraint(
-            ["created_by_email"],
-            ["user.email"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["modified_by_email"],
-            ["user.email"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["organization_id"],
-            ["organization.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_chat_message_chat_id"), "chat_message", ["chat_id"], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_chat_message_chat_id"), table_name="chat_message")
-    op.drop_table("chat_message")
-    op.drop_table("chat")
-    op.execute("DROP TYPE chatrole")
-    # ### end Alembic commands ###
