--- backend/alembic/versions/e0af9e9a55fe_add_sync_status_column.py
@@ -1,43 +1,0 @@
-"""Add sync status column
-
-Revision ID: e0af9e9a55fe
-Revises: a734ec1886d4
-Create Date: 2025-01-23 12:38:02.823505
-
-"""
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "e0af9e9a55fe"
-down_revision = "a734ec1886d4"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # First create the enum type in the database
-    op.execute("CREATE TYPE syncstatus AS ENUM ('ACTIVE', 'INACTIVE', 'ERROR')")
-
-    # Then add the column
-    op.add_column(
-        "sync",
-        sa.Column(
-            "status",
-            sa.Enum("ACTIVE", "INACTIVE", "ERROR", name="syncstatus"),
-            nullable=False,
-            server_default="ACTIVE",
-        ),
-    )
-
-    # Update existing rows
-    op.execute("UPDATE sync SET status = 'ACTIVE'")
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("sync", "status")
-    op.execute("DROP TYPE syncstatus")
-    # ### end Alembic commands ###
